For gcc10 the call to rungakutta results in a type incompatibility.  The integer value -1 is used for particle diameter and particle density which are expecting floats.  These are changed to -1.0 and -1.0 in the function call.
For gcc10 the use of mod(a,b) where b=0 generates an arithmetic error at runtime.  In this case restartwrite is conditioned on mod(time_t,restartwritefreq) with restartwritefreq=0 being a possibility (which presumably implies do not write a restart file).  The conditional is wrapped in acheck to make sure restartwritefreq > 0